libvirt: Add properties to set advanced QXL video RAM settings
This patch adds the following properties:

  - hw_video_ram_bar1      (sets libvirt's 'ram'    XML attribute)
  - hw_video_ram_bar2_ext  (sets libvirt's 'vram64' XML attribute)
  - hw_video_ram_vga       (sets libvirt's 'vgamem' XML attribute)

These properties can be used to specify the video RAM to provide for the
primary bar, the secondary bar extension (64 bit) and the VGA
framebuffer of a QXL graphics adapter.

All three properties are limited by the following extra_specs entries
for a flavor:

  - hw_video:ram_bar1_max_mb
  - hw_video:ram_bar2_ext_max_mb
  - hw_video:ram_vga_max_mb

Note that video RAM values for all three properties and limits are
specified in MB. If a property from above is used, make sure that its
corresponding limit is set in a referenced flavor; otherwise, a runtime
error occurs during instance creation.

Change-Id: If2a3e1686cb13af0e9e54f7aa79dd1696a135c91