libvirt: Add configuration options to set SPICE compression settings
This patch adds the following SPICE-related options to a Nova
configuration:

  - image_compression
  - jpeg_compression
  - zlib_compression
  - playback_compression
  - streaming_mode

These options can be used to enable and set the SPICE compression
settings for libvirt (QEMU/KVM) provisioned instances. Note that those
options are only taken into account if SPICE support is enabled (and
the VNC support is disabled).

Change-Id: Ia7efeb1b1a04504721e1a5bdd1b5fa7a87cdb810